+++ Krowi_AchievementFilter/Objects/Event.lua
@@ -1,0 +1,26 @@
+-- [[ Namespaces ]] --
+local _, addon = ...;
+local diagnostics = addon.Diagnostics;
+local objects = addon.Objects;
+objects.Event = {};
+local event = objects.Event;
+
+-- [[ Constructors ]] --
+event.__index = event;
+function event:New(id, icon) -- Creates a new event
+    local self = {};
+    setmetatable(self, event);
+
+    self.ID = id or 0;
+    self.Icon = icon or 0;
+
+    return self;
+end
+
+function event:AddAchievement(achievement) -- Adds an achievement to the achievement category
+    if self.Achievements == nil then
+        self.Achievements = {}; -- By creating the achievements table here we reduce memory usage because not every category has achievements
+    end
+    tinsert(self.Achievements, achievement);
+    return achievement;
+end