+++ Krowi_Databases/DbManager/DbManagerWPF/ViewModel/ExportViewModel.cs
@@ -22,6 +22,8 @@ namespace DbManagerWPF.ViewModel
 
         public ICommand ExportPetBattlesCommand => new CommandHandler(() => ExportPetBattles(), () => true);
 
+        public ICommand ExportEventsCommand => new CommandHandler(() => ExportEvents(), () => true);
+
         public ICommand ExportAllCommand => new CommandHandler(() => ExportAll(), () => true);
         #endregion
 
@@ -276,7 +278,7 @@ namespace DbManagerWPF.ViewModel
                     if (prevUIMap != keyValuePair.Key.UIMap)
                         sb.Append("}");
                     sb.Append("}");
-                    sb.AppendLine(";");
+                    sb.AppendLine($"; -- {keyValuePair.Key.UIMap.Name}");
 
                     prevUIMap = keyValuePair.Key.UIMap;
                 }
@@ -323,10 +325,14 @@ namespace DbManagerWPF.ViewModel
                 if (!keyValuePairs.ContainsKey((uiMap, category.Function)))
                     keyValuePairs.Add((uiMap, category.Function), new List<Achievement>());
 
-                keyValuePairs[(uiMap, category.Function)].Add(achievement);
+                if (!keyValuePairs[(uiMap, category.Function)].Contains(achievement))
+                    keyValuePairs[(uiMap, category.Function)].Add(achievement);
             }
             else
-                keyValuePairs[(uiMap, null)].Add(achievement);
+            {
+                if (!keyValuePairs[(uiMap, null)].Contains(achievement))
+                    keyValuePairs[(uiMap, null)].Add(achievement);
+            }
 
             foreach (var child in uiMap.Children)
                 LinkMaps(keyValuePairs, category, child, achievement);
@@ -444,6 +450,115 @@ namespace DbManagerWPF.ViewModel
         }
         #endregion
 
+        #region ExportEvents
+        private void ExportEvents()
+        {
+            SortedDictionary<Event, List<(Category Category, Achievement Achievement)>> keyValuePairs = new();
+            var events = eventDM.GetAll();
+            foreach (var @event in events)
+                keyValuePairs.Add(@event, new List<(Category Category, Achievement Achievement)>());
+
+            var categories = categoryDM.GetAll();
+            foreach (var category in categories)
+                LinkAchievementsToEvents(keyValuePairs, category);
+
+            // At this point we have all data
+
+            var sb = new StringBuilder();
+            sb.AppendLine($"-- [[ Exported at {DateTime.Now:yyyy-MM-dd HH-mm-ss} ]] --");
+            sb.AppendLine($"-- [[ This code is automatically generated as an export from ]] --");
+            sb.AppendLine($"-- [[ an SQLite database and is not meant for manual edit. ]] --");
+            sb.AppendLine("");
+            sb.AppendLine("-- [[ Namespaces ]] --");
+            sb.AppendLine("local _, addon = ...;");
+            sb.AppendLine("local objects = addon.Objects;");
+            sb.AppendLine("local event = objects.Event;");
+            sb.AppendLine("local data = addon.Data;");
+            sb.AppendLine("data.ExportedEvents = {};");
+            sb.AppendLine("local exportedEvents = data.ExportedEvents;");
+            sb.AppendLine("");
+            sb.AppendLine("function exportedEvents.Load(e)");
+            sb.AppendLineTabbed(1, "for i, _ in next, e do");
+            sb.AppendLineTabbed(2, "e[i] = nil;");
+            sb.AppendLineTabbed(1, "end");
+            sb.AppendLine("");
+
+            foreach (var keyValuePair in keyValuePairs)
+            {
+                if (keyValuePair.Value.Any() && keyValuePair.Key.ID > 0)
+                    sb.AppendLineTabbed(1, $"e[{keyValuePair.Key.ID}] = event:New({keyValuePair.Key.ID}, {keyValuePair.Key.Icon}); -- {keyValuePair.Key.Title}");
+
+            }
+
+            sb.AppendLine("end");
+            sb.AppendLine("");
+            //sb.AppendLine("local function AddA(e, a)");
+            //sb.AppendLineTabbed(1, "e:AddAchievement(a);");
+            //sb.AppendLine("end");
+            //sb.AppendLine("");
+
+            //sb.AppendLine("function exportedEvents.LoadAchievements(e, a)");
+            //sb.AppendLineTabbed(1, $"if e == nil or (type(e) == \"table\" and #e == 0) then");
+            //sb.AppendLineTabbed(2, $"exportedEvents.Load(e);");
+            //sb.AppendLineTabbed(1, $"end");
+            //sb.AppendLine("");
+
+            //foreach (var keyValuePair in keyValuePairs)
+            //{
+            //    if (keyValuePair.Value.Any() && keyValuePair.Key.ID > 0)
+            //    {
+            //        foreach (var achievement in keyValuePair.Value)
+            //            sb.AppendLineTabbed(1, $"AddA(e[{keyValuePair.Key.ID}], a[{achievement.ID}]); -- {achievement.Name}");
+            //    }
+
+            //}
+
+            sb.AppendLine("function exportedEvents.LoadCategories(e, a)");
+            var appendString = "if ";
+            foreach (var keyValuePair in keyValuePairs)
+                appendString += $"e[{keyValuePair.Key.ID}] == nil or ";
+            appendString = Trim(appendString, " or ");
+            appendString += " then";
+            sb.AppendLineTabbed(1, appendString);
+            sb.AppendLineTabbed(2, $"exportedEvents.Load(e);");
+            sb.AppendLineTabbed(1, $"end");
+            sb.AppendLine("");
+
+            foreach (var keyValuePair in keyValuePairs)
+            {
+                if (keyValuePair.Value.Any() && keyValuePair.Key.ID > 0)
+                {
+                    sb.AppendLineTabbed(1, $"e[{keyValuePair.Key.ID}].Category = a[{keyValuePair.Value[0].Achievement.ID}].Category; -- {keyValuePair.Value[0].Category.Name}");
+                }
+
+            }
+
+            sb.AppendLine("end");
+
+            using var file = new StreamWriter(@"../../../../../../Krowi_AchievementFilter/Data/ExportedEvents.lua");
+            file.WriteLine(sb.ToString());
+        }
+
+        private void LinkAchievementsToEvents(SortedDictionary<Event, List<(Category Category, Achievement Achievement)>> keyValuePairs, Category category)
+        {
+            var categoryEvents = category.GetEvents(true);
+            var achievements = category.GetAchievements();
+
+            foreach (var achievement in achievements)
+            {
+                foreach (var categoryEvent in categoryEvents)
+                    keyValuePairs[categoryEvent].Add((category, achievement));
+
+                var achievementEvents = achievement.GetEvents(true);
+                foreach (var achievementEvent in achievementEvents)
+                    keyValuePairs[achievementEvent].Add((category, achievement));
+            }
+
+            foreach (var child in category.Children)
+                LinkAchievementsToEvents(keyValuePairs, child);
+        }
+        #endregion
+
         private void ExportAll()
         {
             ExportAchievements();
@@ -455,8 +570,13 @@ namespace DbManagerWPF.ViewModel
 
         private string TrimNils(string input)
         {
-            while (input.EndsWith(", nil"))
-                input = input[0..^5];
+            return Trim(input, ", nil");
+        }
+
+        private string Trim(string input, string trim)
+        {
+            while (input.EndsWith(trim))
+                input = input[0..^trim.Length];
 
             return input;
         }